OCE-6371
updates to make the CLI a bit easier to use with a bunch of defaults so you don't have to pass in unnecessary cli parameters.

Also extended it a bit to allo publish events from a CUSTDB and the CUSTDB bits can be passed in via the CLI. Still using a single grab everything SQL statement.

More work required to finish the getMessages CLI which could help the testers. Right now there are broken assumptions in the code, will return to this after those broken assumptions are validated.